+++ core/src/main/java/zemeckis/OnIdleExecutor.java
@@ -1,6 +1,6 @@
 package zemeckis;
 
-import elemental2.dom.DomGlobal;
+import akasha.Global;
 
 /**
  * Run tasks in next Idle callbacks.
@@ -11,6 +11,6 @@ final class OnIdleExecutor
   @Override
   void scheduleForActivation()
   {
-    DomGlobal.requestIdleCallback( deadline -> context().activate( () -> executeTasks( () -> (int) deadline.timeRemaining() ) ) );
+    Global.requestIdleCallback( deadline -> context().activate( () -> executeTasks( deadline::timeRemaining ) ) );
   }
 }